Replace scikit-build-core / CMake with setuptools
Using scikit-build-core and CMake was overkill for copying two files
into `site-packages`. This change replaces them with setuptools, and
also modifies some of the package names to use a dash (i.e.
`numba-cuda`) which I think is more "standard" in the places in which
I've done it here.

Manual testing suggests that the packages built before and after this
change are identical, with a couple of small exceptions:

- The `LICENSE` file is placed in the root by setuptools instead of a
  folder called `licenses`.
- Metadata differs slightly (`WHEEL`, `RECORD`, `METADATA`) due to them
  being generated slightly differently between the build tools.
- setuptools also includes a `top_level.txt` file.

Many thanks to @vyasr for pointing me to the solution to copy files to
`site-packages` with setuptools, as used in
[`rapids-dask-dependency`](https://github.com/rapidsai/rapids-dask-dependency/).