Build and upload sdist
Switching to using the `build` module from `pip` to build the sdist and
to build the wheel from the sdist - this ensures that a wheel is
buildable from the sdist. The use of the `wheel` module to build the
wheel was inherited from other scripts in RAPIDS that were copied, and I
don't think there was any reason to prefer it over the `build` module in
the context of numba-cuda.

CI scripts have been updated accordingly to also upload the sdist (so it
can be inspected if necessary). The `gh-action-pypi-publish` step is
unchanged as it is expected to pick up the sdist from the `dist/`
directory.

The version is bumped to 0.0.15 so that a new version with the sdist can
be published to PyPI once it is merged and tagged.