Fix choice of conda package for includes (on Windows)
The path to CUDA includes may exist in a conda environment even if they
are not present, particularly on Windows where the include path is just
`$CONDA_PREFIX\Library\include`. Therefore, the presence of the include
dir alone doesn't signify that the CUDA includes are actually present in
that location - we also need to check whether a relevant header file
exists.