Bump version to 0.1.0
This increments the minor version rather than the patch version. From
now on, each new release will generally increment the minor version, so
that patch releases are possible for minor versions without introducing
a fourth version number component, as we had to do with v0.0.17.1
(because cuDF was pinned to numba-cuda < 0.0.18).