Make block, thread, and warp indices unsigned
This ports numba/numba#6112 to numba-cuda, as outlined in #109.

Note that for this patch, we don't change the type of `grid()` and
`gridsize()` because these need to be 64 bit (as discovered in
numba/numba#9229 and fixed in numba/numba#9235).

We need to patch the `as_dtype()` function, which is a little
unfortunate, but there's no API for extending its behaviour at present.