Add stats tracking to NRT and tests
This commit adds the `NRT_MemSys` object and implements tracking statistics for
allocations and frees. Initialization of a singleton MemSys object for an
entire GPU is added along with setup for each kernel to point its MemSys
pointer to the global MemSys object.

Tests for stats tracking are ported from the CPU target, with some additional
mock NumPy functions required by the tests.

Additional configuration variables are used by this functionality, so
`_readenv` is moved into `numba.cuda.utils`.