Implement debuginfo bool name fix (numba/numba#9888) in numba-cuda (#106)
As a workaround until numba/numba#9888 is merged and available in an
upstream Numba release, we interject our own modified bytecode
translation pass that names bools correctly (`$bool<N>` as oppoosed to
`bool<N>`).

To avoid duplicating the whole untyped pipeline definition, we take the
definition from upstream Numba and modify it to use our pass.

Add a test for the bool name fix.

---------

Co-authored-by: jiel-nv <jiel@nvidia.com>