Fix the debug info of GridGroup type
The customized CUDA front end type 'GridGroup' maps to the back end
type 'ir.IntType(64)' which is defined in numba.cuda.models.

The numba DIBuilder emit variable types for simple numeric types by
looking into back end type which falls through the 'DW_ATE_float'
because the 'datamodel.fe_type' is not 'types.Integer'.

So emitting this type through the CUDADIBuilder is needed. Since the
type represents the handle of a group object, giving it the encoding
of 'DW_ATE_unsigned' should be reasonable.

A test is also added in this change.

This change addresses https://github.com/NVIDIA/numba-cuda/issues/107