CUDASIM: Support detection of device function calls
The hardware target has supported detection of whether a decorated
function is used as a kernel or device function for a long time.
However, the simulator has needed to be informed via the `device=`
kwarg.

This changes the simulator to determine when a decorator is used as a
device function - it can do this by checking for an existing "kernel
context" when a decorated function is called - if one exists, it must be
being used as a device function.