Add Bfloat16 Low++ Bindings (#166)
This PR introduces low++ bfloat16 bindings to Numba-CUDA. They are exposed via `numba.cuda.cuda_bf16` module. Available intrinsic functions includes:

* htrunc
* hceil
* hfloor
* hrint
* hsqrt
* hrsqrt
* hrcp
* hlog
* hlog2
* hlog10
* hcos
* hsin
* hexp
* hexp2
* hexp10

This also vendors bfloat16 headers for CUDA 11 and 12 and float16 headers with separate versions for 11 and 12 (previously only the CUDA 11 float16 headers were vendored).

`StringIO` and `BytesIO` objects can be used in linkable code, in order to defer the generation of the data for a linkable code object until it is actually used / needed.

A workaround for numba/numba#10047, "fix `IntEnumMember` conversion" is needed - the issue is unrelated to this PR but is triggered by it. This is a temporary workaround until the Numba PR is merged and available in released and supported versions.

---------
Co-authored-by: Graham Markall <gmarkall@nvidia.com>