CodeLibrary: Add linking files to library when linking another library
When calling a device function that itself called an external function
implemented inside linkable code, the link would fail because the
linkable code was never added to the link for the calling kernel.

This commit fixes the issue by adding the linked library's list of
linking files to the current code library when it is linked - this
allows the external code dependencies of all called functions to
propagate up to the linking dependencies of the outermost function.