Tidy up / refactor and document declare_device_function
We reduce the number of interfaces here (`declare_device_function` vs
`declare_device_function_template`) and use the standard
`make_concrete_template` function to generate a typing template instead
of inventing our own `device_function_template` class for the typing.

The `link` property of an `ExternFunction` is no longer needed - it was
previously required when the `Dispatcher` had to determine the list of
objects to link; it is now removed.