Add external linked code to code libraries
This introduces a new code library, an `ExternalCodeLibrary`. It can
only hold linking dependencies, so it is not possible to add IR modules
to it or get functions or generated code out of it.

It is also not serializable, like other code libraries - this could be
added, but has not been done at this point because serializing code
libraries with external dependencies is already unsupported.

When declaring an external function with `declare_device()`, any linked
libraries are added to a new `ExternalCodeLibrary` object, which is then
added to the list of libraries for the function in the target context.
Core Numba functionality then ensures that the external code library is
added to the linking dependencies of any callers.

With linking dependencies now being tracked through code libraries,
there is no need for the `Dispatcher` to try and pull together a list of
items to link to a kernel. The implementation of this process was
already broken, as it failed to see through functions implemented using
the high-level extension API (`@overload`, etc.). It is now removed.