Add CUDA simulator test workflow
The simulator needs to be tested with Python 3.12 - it presently does
not fully work on Python 3.13 due to the change in behaviour for
`PyObject_HasAttr()` to no longer silently ignore all errors as it did
previously.

Numba's typecode fingerprinting calls something which is approximately
`PyObject_HasAttr("_numba_type_")` on NumPy arrays during some execution
paths of the simulator, which causes NumPy to raise `ValueError("no
field of name _numba_type_")`. Since Python 3.13 this causes masses of
warnings to be raised that flood the output.

References:

- Python 3.12 `PyObject_HasAttr()`:
  https://docs.python.org/3.12/c-api/object.html#c.PyObject_HasAttr
  > "Exceptions that occur when this calls __getattr__() and
  > __getattribute__() methods are silently ignored."
- Python 3.13 `PyObject_HasAttr()`:
  https://docs.python.org/3.13/c-api/object.html#c.PyObject_HasAttr
  > "Exceptions that occur when this calls __getattr__() and
  > __getattribute__() methods aren’t propagated, but instead given to
  > sys.unraisablehook(). For proper error handling, use
  > PyObject_HasAttrWithError(), ..."
- CPython commit adding the recommended functions:
  https://github.com/python/cpython/pull/109025
- CPython issue: https://github.com/python/cpython/issues/108511