Allow External Code to Use Cooperative Group (#240)
Currently, Numba-CUDA detects the use of `cudaCGIntrinsicHandle` in PTX and auto launch the kernel with `cuLaunchKernelCooperative`. This only works when the Numba kernel itself uses the internal typing extensions which guarantees the use of this call. For external device functions that invoke CG intrinsics, they don't have a way to pass in info to inform Numba to cooperative launch kernel. This PR adds that support.

* Add cooperative attr to codelib
* make sure that declare_device can specify use of cooperative launch and is traced in codelib
* update declare device usage and fix type
* add external coop kernel launch example
* avoid searching the ptx to determine if cooperative launch is needed
* Improve cooperative launch test link usage
* persist `use_cooperative` for codelib
* create a separate cg test group to avoid vending run_in_separate_process
* Update numba_cuda/numba/cuda/tests/cudapy/test_caching.py

---------

Co-authored-by: Graham Markall <535640+gmarkall@users.noreply.github.com>