Add debug support for dynamic types
This change implements debugging support for polymorphic variables via a C Union
approach.

(1) A pre-lower scan on function body checks all the assignment instructions for
versioned / unversioned  targets names and generates a polymorphic types map for
each qualified source name;
(2) Frond end Literal types are deduplicated before recorded into an entry.
(3) Leverage the UnionType and UnionModel to create and handle the aggreate type
where the "payload" of field representing data types in memory.
(4) The allccation of memory uses a single element aggregate type and necessary
type casts are made for easy retrieval of member fields of the artificial union.

This fulfills the RFE in nvbug#5267572.