Skip compute capability 10.3 on CUDA 12.8
The compute capability 10.3 was added with CUDA 12.9. However the logic
here treats the whole range of CUDA 12.8 from 5.0 to 12.0 as inclusive.
As a result 10.3 was included with CUDA 12.8 where it is not supported.

To fix this allow each CUDA version to take multiple ranges. That way
two ranges can be used for CUDA 12.8. First a range including 5.0 to
10.1. Second a range including 12.0 only. That way CUDA 12.8 can be
written such that it excludes 10.3. The rest of the CUDA versions simply
use one range for their compute capabilities.