Add CUDA 12.9 and new architectures (#273)
* Add CUDA 12.9 and new architectures

* Skip compute capability 10.3 on CUDA 12.8

The compute capability 10.3 was added with CUDA 12.9. However the logic
here treats the whole range of CUDA 12.8 from 5.0 to 12.0 as inclusive.
As a result 10.3 was included with CUDA 12.8 where it is not supported.

To fix this allow each CUDA version to take multiple ranges. That way
two ranges can be used for CUDA 12.8. First a range including 5.0 to
10.1. Second a range including 12.0 only. That way CUDA 12.8 can be
written such that it excludes 10.3. The rest of the CUDA versions simply
use one range for their compute capabilities.

* Fix `list`-comprehension ordering

* Use 10.1 as upper bound in CUDA 12.8

Co-authored-by: Graham Markall <535640+gmarkall@users.noreply.github.com>

* Run `ruff format` to fix CI check-style

* updated ci imgs

* updated ci imgs take2

* updated ci imgs take3

---------

Co-authored-by: Graham Markall <535640+gmarkall@users.noreply.github.com>
Co-authored-by: Keith Kraus <keith.j.kraus@gmail.com>