Make 7.5 the default compute capability (#275)
* Update supported toolkit and CC versions

- 11.1 is not supported; it is removed.
- 12.9 added; additional support for CC 12.1.

* Ensure CC 7.5 or greater used by default

The CUDA programming guide mentions that the next major release will
remove compilation and library support prior compute capabilities.

CC 7.5 is now used by default (unless the user has specified a greater
CC as the default using a config variable). This is safe as all
currently-supported CUDA toolkits support CC 7.5.

* Update tests for default CC 7.5

Most of these tests specified a particular CC under the assumption that
the minimum was 5.0 or 5.2 - now it is 7.5, the minimum no longer needs
specifying.

A couple of other test cases that checked for different behaviour across
CCs are also handled.

* Use LOWEST_CURRENT_CC in test_nvvm_compile_nullary_option