Add support for launch bounds (#272)
Adds support for launch bounds, specified as the `launch_bounds` kwarg of the `@cuda.jit` decorator or `cuda.compile()` function. The API design is intended to follow that of the [Launch Bounds API in CUDA C/C++](https://docs.nvidia.com/cuda/archive/12.9.0/cuda-c-programming-guide/index.html#launch-bounds).

Bounds may be specified as a scalar or a tuple of between one and three items. Tuple items provide:

- The maximum number of threads per block,
- The minimum number of blocks per SM,
- The maximum number of blocks per cluster.

If a scalar is provided instead, it is used as the maximum number of threads per block.

A sketch of some possile uses of the API in the `@cuda.jit` decorator:

```python
# Specify max threads per block only:
@cuda.jit(launch_bounds=128)
def kernel(...):
    ...

# Specify max threads per block and minimum blocks per SM:
@cuda.jit(launch_bounds=(max_threads_per_block, min_blocks_per_sm))
def kernel(...):
    ...

# Specify all three kinds of bounds:
@cuda.jit(launch_bounds=(max_threads_per_block, min_blocks_per_sm, max_blocks_per_cluster))
def kernel(...):
    ...
```