Fix lineinfo generation when compile_internal used (#271)
Lineinfo generation is broken by function implementations that generate
code via `context.compile_internal()`. The root cause is that the
implementation eventually reaches upstream Numba's
`BaseContext._compile_subroutine_no_cache()` method, which ignores the
flags in the context stack and creates new ones. The outcome of this is
that the debug info kind is forgotten, leading to a default debug info
kind of `"FullDebug"` being emitted, which then enables the PTX `debug`
target, leading to deoptimized code.

This change works around the issue (pending a fix upstream) by
overriding the `_compile_subroutine_no_cache()` implementation to
use flags from the context stack when they are otherwise not provided.
The fix to upstream will look like a similar modification of the method.

The `CUDAFlags` class is moved to its own module to avoid circular
import dependencies between `compiler.py` and `target.py`.

Fixes #271.