Use NVRTC for supported CCs
Changes in the compute capability support matrix in nvvm.py will
continue to be needed with new CUDA versions if we maintain a list of
explicitly-supported compute capabilities. NVRTC supports retrieving the
supported list programmatically, so we switch to using it instead.

This does assume that the user's environment has a consistent set of
components (NVVM, NVRTC, etc.) - this is generally expected to be the
case with recent developments in package management, and there's little
we can do about an inconsistent environment anyway.

Changes outside of nvvm.py / nvrtc.py are to accommodate the movement of
this functionality. A major side effect is that we no longer need to
initialize the list of supported CCs prior to forking, because we don't
need to use the CUDA runtime to populate the supported CC list.