Fix `DeviceNDArray.device_ctypes_pointer` to always be ctypes (#295)
When the NVIDIA bindings are in use, `device_ctypes_pointer` should
always be a ctypes pointer, otherwise the interface changes depending on
which set of bindings are in use. This has been a long-standing
inconsistency, but it needs to be resolved.

The fundamental change is only to `DeviceNDArray.device_ctypes_pointer`.
The associated changes are all fixups to maintain internal consistency.