Improve the debug mode code generation for dynamic types.
The polymorphic variables do not need zero finalization after the usage of each
variation, since they share the same memory in the same artificial debug union.
The cleanup work is done in the last basic block of the function.

Also, do not use the cached temporaries as the pointer for the variations, they
need to be regenerated and type casted from the debug union on each load/store.