Don't skip debug tests when LTO enabled by default
These used to be skipped when LTO was enabled because it was
unconditionally enabled, even for kernels where `debug=True`. Now if
`debug=True` LTO is disabled, so it is not a problem to run the debug
tests even when LTO is enabled.

This fixes something of a regression in our test coverage, where these
tests were run in fewer settings due to LTO being enabled by default.