Use NVRTC to get runtime version
We only used the CUDA runtime library to get the runtime version so that
we could populate the list of supported compute capabilities in nvvm.py.
Now that we don't do this, and that NVRTC provides the CUDA toolkit
version, there is no need to use the CUDA runtime API at all.

The Numba API for the runtime version is not deleted in case it was used
by external code - instead, it uses NVRTC to obtain the toolkit version.

Because NVRTC used the runtime version to determine what prototypes to
bind, we need to stop doing that to avoid a circular dependency /
deadlock - instead of checking the runtime version and creating the list
of prototypes, we try to add all known prototypes, and ignore errors in
those related to LTOIR, which can occur with CUDA 11 where they were not
present.

The `runtime.is_supported_version()` API and its test is removed - it
would always have been `False` on CUDA 12 (incorrectly) and this has
never been reported as an issue, so it seems very unlikely that anyone
was using it.