Default to at least compute capability 7.5
The change to use NVRTC for the supported compute capabilities also had
the implicit effect of making the default compute capability the lowest
supported by the installed NVRTC version. We need it to default to at
least 7.5 (unless specified higher by the user) to maintain the
behaviour of the compute capability logic from nvvm.py that was
replaced.