Replace Numbast generated lower_casts
cuda_fp16.h header in CUDA 11 and CUDA 12 differs in that the cu11
version does not define a number of conversion functions. For example
`operator char` is present in cu12 but not in cu11.

This is a problem because if bindings are generated using cu11 headers,
then they do not contain the casting functions and if they are generated
using cu12 headers, they call functions that don't exist in cu11 causing
the operations to fail for cu11. Since numba-cuda originally supported
these functions for cu11, dropping support is not acceptable to avoid
breaking pre-existing code.

To work around this issue, we generate bindings using cu12 headers
and delete all the numbast generated conversion functions. These are
replaced by custom logic that uses PTX that's supported in both cu11
and cu12.