`MmemoryPointer` ensure `CUdeviceptr` used with NVIDIA binding
The docs state that `pointer` should be a `c_void_p`, but the rest of
the driver binding code expects the pointer to be represented as a
`CUdeviceptr` when the NVIDIA bindings are in use.

If we're given a `c_void_p` by the user when using the NVIDIA bindings,
we can work around this by converting it to a `CUdeviceptr`.

This was first noticed when testing PyArrow with Numba-CUDA. See
https://github.com/apache/arrow/issues/47128