[testing] Add filecheck test infrastructure
Adding pattern-based checking infrastructure allows us to describe
the IR and assembly expected under different conditions and expressed
right next to the code that produced it. This is closer to the LLVM
testing standards, and will help is iterate more quickly and add test
coverage while we are in the process of vendoring in components
from numba core.

This is a request for comment, as filecheck tests will be meaningfully
different from existing numba-cuda tests.