[RFC,TESTING] Add filecheck test infrastructure (#342)
Adding pattern-based checking infrastructure allows us to describe the IR and assembly expected under different conditions right next to the code that produces it. This is closer to the LLVM testing standards, and will help is iterate more quickly and add test coverage while we are in the process of vendoring in the components from numba core we need to modify in a cuda-specific way.

This PR changes a couple tests to use the filecheck patterns. For example, in numba-cuda today there are tests like `self.assertIn('function_name', llvm_ir)`. Now, we can express the same check in the docstring of the kernel, and all the checks we would like to perform may live inside the kernel they come from.

Note also that the filecheck that has been added to the list of dependencies is _not_ the filecheck that is compiled as a part of the LLVM project, but [is an attempt to replicate the functionality in a pure python package](https://github.com/AntonLydike/filecheck/tree/main).

* [testing] Add filecheck test infrastructure
* Install filecheck in conda CI scripts
* Add CUDADispatcher dummy class for simulator

---------
Co-authored-by: Graham Markall <gmarkall@nvidia.com>