Add more debuginfo tests
Additional debug info tests were adapted from Numba core
debuginfo tests, but using filecheck. For example, the
original tests searched the IR with regular expressions
and built up a dictionary of debuginfo values and the
types they correspond to - now we use filecheck capture
variables the first time they are encountered, and we refer
back to them later in the checks.