[Refactor] Vendor in Typing Templates
The Signature class is also modified to inherit from the upstream
`numba.core.typing.templates.Signature` to pass subclass checks in the upstream
`AbstractTemplate` that is still used by parts of upstream that haven't
been vendored in yet. Likewise, there are calls to the vendored-in `apply` that pass in the upstream Signature so the subclass check in the vendored-in AbstractTemplate is extended to check for both.

These are hacks marked as such, so we can undo them as soon as we can.