Split out debuginfo tests for types
The basic scalar and array types changed slightly in llvmlite
45; Jie submitted some fixes to eliminate redundant types,
which causes some differences in the IR we generate.

This PR addresses the issue in two ways:
- For the scalar types, each type is broken out into its own
	test case. This removes the differences we see in llvmlite45,
	because there are no common debuginfo metadata nodes that
	can be eliminated in the first place.
- For array types (float32 is the only one tested this way in both
	upstream and numba-cuda) I added a check prefix for
	llvmlite version >=45 and <=44. The IR is just different
	enough to make it difficult for the same checks to work
	with both versions. After Jie's fix, llvmlite no longer
	creates redundant debuginfo metadata nodes for basic scalar
	types.

Although it takes up quite a bit of space, I find it easier to navigate
the checks with IR snippets in the test code than reading regular
expressions. This is a matter of taste though, so I'm open to
suggestions.

Note the use of subtests; this should be removed in favor of
parameterized tests with pytests once we are fully moved over.

This might also be a good prompt for us to start testing with
newer/development versions of llvmlite.