[Refactor] Vendor in Typing Templates (#360)
The Signature class is also modified to inherit from the upstream
`numba.core.typing.templates.Signature` to pass subclass checks in the
upstream `AbstractTemplate` that is still used by parts of upstream that
haven't been vendored in yet. Likewise, there are calls to the
vendored-in `apply` for `AbstractTemplate` that pass in the upstream
`Signature` so the subclass check in the vendored-in `AbstractTemplate`
is extended to check for either.

These are hacks marked as such, so we can undo them as soon as we can.
Afaict, we can't use subclass checks/hooks since this would require
changing the upstream Signature class.

---------

Co-authored-by: Graham Markall <gmarkall@nvidia.com>