Vendor in Bytecode Interpreter (#355)
I only vendored-in the interpreter with Python > 3.9 because:
* The main branches of Numba core and llvmlite do not support Python <=
3.9, and there are lots of checks for Python versions in this source
file.
* I would like the first check-in of the interpreter to be unchanged, so
that the original version exists in our git history in case we need to
track down bugs introduced in our interpreter with CUDA-specific
changes.

This is why the original Numba interpreter is still imported in 3.9.
If we like, I will go through the version-specific parts of the
interpreter as a follow-up, and try to edit the interpreter such that we
can still use it with older pythons.
I would like to do that in a second step to be sure I do not introduce
any bugs in the vendoring-in process.

Code coverage reported 46% from this source file after running our test
suite. There are a few reasons.
* Much of the interpreter covers features that I don't believe we
support, namely dictionaries and exceptions (please correct me if I'm
wrong)
* Some of it is gated behind Python versions (eg a different method for
3.10 vs 3.11)
* I do not see any tests in Numba checking the behavior of specific
opcodes - they appear to be integrated tests where the kernel's behavior
and IR are inspected, but the opcodes used by the interpreter are not.

The interpreter was pulled in from numba commit 262b49a3e