Fix correctness issue in debug mode code generation for dynamic types. (#305)
The polymorphic variables do not need zero finalization after the usage
of each variation, since they share the same memory in the same
artificial debug union. The cleanup work is done in the last basic block
of the function.

Also, do not use the cached temporaries as the pointer for the
variations, they need to be regenerated and type casted from the debug
union on each load/store.

The correctness test added in this patch produces wrong result with
existing code, as below,
[0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0]
with this patch, the result will be produced correctly as expected,
[1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0]

---------

Co-authored-by: Graham Markall <gmarkall@nvidia.com>