Port tests to use numba cuda
Test test_unhandled_undefined was skipped because it relies on
potentially undefined return values and we do not support return values
at all.

Test test_issue5482_objmode_expr_null_lowering was skipped because the
object mode is not supported for the numba-cuda target.

Several tests that used custom compilation pipelines were omitted
because custom pipelines do not work with numba.cuda.jit. We may want to
re-add these tests if we add this functionality, but I did not include
custom pipelines in this PR to keep the focus on vendoring in this
component and not feature work.