Improve Bfloat16 Support in Numba-CUDA (#376)
This is an overhaul of the bfloat16 support. The PR improves bfloat16 by
implementing a unified type object to the constructor handle, defining
proper cast and unify rules of bfloat16 to other numeric types, adding
*initial* host side support of this type (leveraging ml_dtypes package),
printing support in kernel, as well as plethora of intrinsics.

Details:
1. You may now `print(x)` on device.
2. You may now use comparison intrinsics, such as heq, hmax, hmax_nan
etc. on device.
3. You may also use conversion intrinsics, such as
`int64_to_bfloat16_rz` etc. on device; bit cast intrinsics, such as
`int16_as_bfloat16` on device.

---------

Co-authored-by: Michael Wang <isVoid@users.noreply.github.com>