Vendor in postproc module for CUDA-specific refactoring (#418)
This module has few dependencies and no standalone tests, so it has been
vendored in as-is (with the copyright modifications).

A few unrelated imports were importing modules from upstream numba that
we have already vendored in, so I corrected them to use our copy. If we
would like this PR to be strictly related to the postproc module, I can
revert these changes.