Vendor in SSA for CUDA-specific refactoring (#417)
The tests had to undergo some significant refactoring - I tried to
preserve as much of the original test coverage as I could, but here are
notes on some tests that I had to remove entirely:
* Test test_issue5482_objmode_expr_null_lowering was skipped because the
object mode is not supported for the numba-cuda target.
* Several tests that used custom compilation pipelines were omitted
because custom pipelines do not work with numba.cuda.jit. We may want to
re-add these tests if we add this functionality, but I did not include
custom pipelines in this PR to keep the focus on vendoring in this
component and not feature work.

Otherwise, it was a 1-1 replacement of numba's jit and njit with
numba.cuda.jit, and return values were replaced with assignments to
output arrays.