Add import of optional in `load_additional_registries()`
This is imported as a side effect in the upstream `numbers.py`, causing
registrations to the builtin registry for implementations involving the
`Optional` type.

The side-effecting import could be added back to numbers.py, but I think
the right place to be keeping imports that have registrations as side
effects is in `load_additional_registries()` so that they're all in one
place and we can work through handling each case. In each vendored
module, we should create a new registry and explicitly add it to the
CUDA target context, rather than jamming many items arbitrarily into the
"builtin registry".