[Refactor][NFC] Vendor in testing support utils for CUDA-specific changes (#432)
This change vendors in numba.testing.support, specifically the Mixins
that we use for testing for memory leaks and checking warnings.

Notes from review: It's a little non-obvious why the change to the simulator's
`RTSys` was needed. It is because the `BaseUFuncTest` switched to the
Numba-CUDA `MemoryLeakMixin` which initializes the Numba-CUDA `rtsys`. Prior to
this PR, it was erroneously using the CPU-based `MemoryLeakMixin`.