Vendor the base module for CUDA-specific refactoring (#445)
There are no standalone tests for this module. It is imported from the
cpu and caching modules upstream, but we have not pulled in the cpu
module and the caching module has already been vendored but BaseContext
was not (as it was unused by the upstream caching module anyways).

There are some tests which import this module, but we have versions of
the same tests and they do not use the base module:

* numba/numba/tests/test_target_overloadselector.py
* numba/numba/tests/test_ufuncs.py
* numba/numba/tests/test_target_extension.py

---------

Co-authored-by: Graham Markall <gmarkall@nvidia.com>
Co-authored-by: Graham Markall <535640+gmarkall@users.noreply.github.com>