Avoid Using Device Compute Capability for Linker Class (#429)
When debugging with @gmarkall, we discovered that the linker class is
using the current device's compute capability for cuda source
compilation. Because `compile_ptx` can be used
to compile target code for a different architecture from current device.
This PR pass through the CC used to construct the linker class to the
cuda source compilation. This avoids
CC inconsistencies between compile and linking.

---------

Co-authored-by: Michael Wang <isVoid@users.noreply.github.com>